+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add the prometheus log level option
+
 ### Fixed
 
 - Fixed the way VPA `maxAllowed` parameter for memory is calculated so that we